rename: SurveySupplementScreen->SurveyMedicationOkScreen
